pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rlpe-radiolarian-plate-extractor"
version = "0.1.0"
description = "Local pipeline for extracting radiolarian plates, captions, labels, and taxon names from literature PDFs"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests",
  "numpy",
  "opencv-python",
  "pymupdf",
  "pandas",
  "tqdm",
  "scikit-image",
]

[project.optional-dependencies]
ocr = ["paddleocr", "easyocr"]
taxon = ["taxonerd", "spacy"]
sam = ["sam2"]
gemma = ["transformers", "accelerate", "bitsandbytes", "pillow"]

[project.scripts]
rlpe = "rlpe.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
